requirements.txt
fastapi
uvicorn
streamlit
google-api-python-client
google-auth
requests
